restructure to unittest scaffolding
